+++ tests/specs/keyof/quotedProperty.d.ts
@@ -1,0 +1,5 @@
+interface IntrinsicElements {
+    "var": string
+}
+
+type Keys = keyof IntrinsicElements