+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Transform `Promise` to `JS.Promise` (by @nojaf) ([GH-33](https://github.com/glutinum-org/cli/pull/33))
 * Optional argument of F# Method are prefixed with `?` instead of suffixing them with `option`
+* Sanitize names coming from TypeScript by removing surrounding quotes (`"`, `'`)
 
 ### Added